Feat: Globla refacto
