+++ src/controller/cli/game_controller.h
@@ -1,0 +1,6 @@
+#ifndef GAME_CONTROLLER_H
+#define GAME_CONTROLLER_H
+
+void runGame(const char *scoreFilePath);
+
+#endif // GAME_CONTROLLER_H
